dev: justfile touch-ups
Refines the "clean" target by reusing more of the pnpm-specified clean
actions, and makes running the tests locally a bit more straightforward.

dev: add test benchmark script

+++ justfile
@@ -3,22 +3,27 @@ install:
   pnpm install
 
 # Build the apps via turbo
-build: install
-  pnpm turbo build --cache-dir=.turbo
+build:
+  pnpm turbo build
 
 # Compile the WASM dependencies via turbo
-wasm: install
-  pnpm turbo compile --cache-dir=.turbo
+wasm:
+  pnpm turbo compile
 
 # Compile the WASM dependencies via turbo (alias)
-compile: install
+compile:
   @just wasm
 
 # Remove all cached build artifacts
 clean:
+  pnpm clean
+  pnpm clean:modules
+  pnpm clean:vitest-mjs
   rm -rf .turbo
   fd -t d node_modules --no-ignore -X rm -r
   fd -t d target --no-ignore -X rm -r
+  fd -t f tsconfig.tsbuildinfo --no-ignore -X rm
+  fd -t d flake-inputs --no-ignore --hidden -X rm -r
 
 # Wrapper for all linting targets
 lint:
@@ -26,12 +31,12 @@ lint:
   @just lint-rust
 
 # Lint the turbo resources
-lint-turbo: install
-  pnpm turbo lint:strict --cache-dir=.turbo
+lint-turbo:
+  pnpm turbo lint:strict
 
 # List Rust code
-lint-rust: install
-  pnpm turbo lint:rust --cache-dir=.turbo
+lint-rust:
+  pnpm turbo lint:rust
 
 # Build top-level debug container
 container: clean
@@ -48,13 +53,25 @@ veil-container:
 
 # Configure OS deps for Playwright tests
 install-playwright: install
+  test -n "$IN_NIX_SHELL" && echo "ERROR: playwright deps are already installed in the nix env" && exit 1
   pnpm playwright install --with-deps chromium
 
+# Run all test suites
+test:
+  # Run rust tests
+  @just test-rust
+  # Run turbo playwright tests
+  @just test-turbo
+
 # Run the turbo test suite
-test-turbo: install
-   pnpm turbo test --cache-dir=.turbo
+test-turbo:
+   pnpm turbo test
 
 # Run the Rust test suite
-test-rust: install
-   pnpm turbo test:wasm --cache-dir=.turbo
-   pnpm turbo test:cargo --cache-dir=.turbo
+test-rust:
+   pnpm turbo test:cargo
+   pnpm turbo test:wasm
+
+# Run test suites locally and gather timing information
+benchmark-tests:
+  ./scripts/benchmark-tests
